Required for Cisco IOS oddities
Minor change for Cisco IOS driver. If inline_transfer is True, then the Cisco IOS driver (Netmiko) will use the SSH channel to transfer the configuration change. If the inline_transfer is False (unset), then the Cisco IOS driver will attempt to use SCP.

For some router operating systems, attaining a suitable privilege level is fraught with trouble. Inlining is the easiest way to solve the issue in some cases.

For other drivers (non 'ios'), unless the driver has an internal 'inline_transfer' field, it has no affect. By default is not set and is not required.