inline_transfer addition
This input parameter allows users to set config transfers to either be via SSH or SCP.

Although I'm using the 'TRUE' string here, the parameter in the Cisco IOS is a boolean. Therefore if it's passed to the parameter, it evaluates as True, if it is not set, the parameter is evaluated to be False; it becomes true from a side-effect as opposed to direct manipulation of the boolean. 

This is specifically covering the Cisco IOS driver at this time. Adding this parameter might also be useful for other drivers.